// src/apis/models/FireBytes_Unified_Wave_BuildingFloorModels_Dtos_BuildingFloorModelEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_BuildingFloorModels_Dtos_BuildingFloorModelEditDto = {
    id?: string | null;
    modelName?: string | null;
    fileId?: string | null;
    description?: string | null;
    type: number;
};